chore(dependencies): update package versions and restructure dependencies
This commit updates the versions of several dependencies in the `pnpm-lock.yaml` and `package.json` files, including React and Radix UI components, to ensure compatibility and leverage the latest features. Additionally, it restructures the dependencies by moving `shared-data` back into the `dependencies` section of `package.json`, enhancing the project's organization. These changes improve the overall stability and maintainability of the application.